Unify src/vis/layers.js and src/vis/vis/layers.js
