+++ src/vis/layers.js
@@ -1,4 +1,5 @@
-var Layers = require('./vis/layers');
+var _ = require('underscore');
+var log = require('cdb.log');
 var TileLayer = require('../geo/map/tile-layer');
 var WMSLayer = require('../geo/map/wms-layer');
 var GMapsBaseLayer = require('../geo/map/gmaps-base-layer');
@@ -37,6 +38,28 @@ function transformToHTTPS (tilesTemplate) {
   }
   return tilesTemplate;
 }
+
+// layer factory
+var Layers = {
+
+  _types: {},
+
+  register: function (type, creatorFn) {
+    this._types[type] = creatorFn;
+  },
+
+  create: function (type, data, options) {
+    if (!type) {
+      log.error('creating a layer without type');
+      return null;
+    }
+    var LayerClass = this._types[type.toLowerCase()];
+
+    var layerAttributes = {};
+    _.extend(layerAttributes, data, data.options);
+    return new LayerClass(layerAttributes, options);
+  }
+};
 
 Layers.register('tilejson', function (data, options) {
   var url = data.tiles[0];
