+++ test/spec/vis/vis-view.spec.js
@@ -13,7 +13,6 @@ var VisModel = require('../../../src/vis/vis');
 var VizJSON = require('../../../src/api/vizjson');
 
 require('../../../src/vis/overlays'); // Overlay.register calls
-require('../../../src/vis/layers'); // Layers.register calls
 
 describe('vis/vis-view', function () {
   beforeEach(function () {
